REFACTOR: Rename `AlgebraicWiringDiagrams` submodule.
To `DirectedMonoidalWiringDiagrams`, in preparation for a similar
submodule for UWDs.
