+++ src/wiring_diagrams/WiringDiagrams.jl
@@ -4,7 +4,7 @@ using Reexport
 
 include("Directed.jl")
 include("Undirected.jl")
-include("Algebraic.jl")
+include("MonoidalDirected.jl")
 include("Algorithms.jl")
 include("Expressions.jl")
 include("Serialization.jl")
@@ -13,7 +13,7 @@ include("JSON.jl")
 
 @reexport using .DirectedWiringDiagrams
 @reexport using .UndirectedWiringDiagrams
-@reexport using .AlgebraicWiringDiagrams
+@reexport using .MonoidalDirectedWiringDiagrams
 @reexport using .WiringDiagramAlgorithms
 @reexport using .WiringDiagramExpressions
